Added caching suppot to the download scripts. Also fixed a deploy issue where versioneer would report a dirty version because of the deploy configuration.

+++ scripts/windows_base_install.ps1
@@ -16,8 +16,12 @@ $SCRIPTS_DIR=$PSScriptRoot
 # Downloads pip into $PYTHON_SCRIPTS
 & $PYTHON_EXE "$SCRIPTS_DIR\get-pip.py"
 
-# Installs the Visual Studio 14 (2015) compiler
-& choco install vcbuildtools
+# Installs the Visual Studio 14 (2015) compiler if it doesn't already exist.
+$VS_HOME="C:\Program Files (x86)\Microsoft Visual Studio 14.0\"
+If(!(test-path $VS_HOME))
+{
+  & choco install vcbuildtools
+}
 
 # The default VS2017 direrctory needs to be deleted so that it doesn't interfere with the VS2015 build.
 Remove-Item "C:\Program Files (x86)\Microsoft Visual Studio\2017" -Force -Recurse
